backend/requirements.txt
aiohappyeyeballs==2.5.0
aiohttp==3.11.13
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
attrs==25.1.0
boto3==1.36.2
botocore==1.36.2
certifi==2025.1.31
charset-normalizer==3.4.1
distro==1.9.0
Django==5.1.3
django-cors-headers==4.6.0
django-storages==1.14.4
djangorestframework==3.15.2
dotsi==0.0.3
frozenlist==1.5.0
geographiclib==2.0
geopy==2.4.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.9.0
jmespath==1.0.1
multidict==6.1.0
openai==1.66.2
propcache==0.3.0
pydantic==2.10.6
pydantic_core==2.27.2
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
requests==2.32.3
s3transfer==0.11.1
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.2
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
yarl==1.18.3
